successfully added remaining balance to the card users page

+++ database/migrations/2018_06_19_073534_create_users_table.php
@@ -22,7 +22,8 @@ class CreateUsersTable extends Migration
             $table->string('address');
             $table->integer('home_phone');
             $table->integer('mobile_phone');
-            $table->decimal('balance', 8, 2);
+            $table->integer('balance');
+            //$table->decimal('balance', 8, 2);
             $table->timestamps();
         });
     }
